Update common submodule
In particular pick this:

  commit 93a7f3af5c23ece6a8e092827ed5928a8973fd3c
  Author: Richard W.M. Jones <rjones@redhat.com>
  Date:   Wed Apr 24 12:08:01 2024 +0100

    options: Allow nbd+unix:// URIs

(cherry picked from commit 7968de46f17eed7f18c200c6528ebdfec5c3f279)

RWMJ: Although this is strictly a new feature, we don't have much
choice except to pick it for the stable 1.52 branch since it's part of
the submodule.  However it's probably a good idea to have this as it
aligns with our goal to have better support NBD URIs everywhere.