build: Only require augeas and hivex if also building the daemon
These libraries are only needed if building the daemon (so not if
using ./configure --disable-daemon).  Move the tests so we only check
for these libraries if the daemon is enabled in the build.

The daemon also requires ocaml-augeas (it uses both the augeas C
library directly and the OCaml bindings for augeas, for reasons).
However that test is already disabled if the daemon is not being
built.

Reported-by: Mohamed Akram
Fixes: commit dfa9dee775acff445de0f626dc54088775e2150e
Fixes: commit 8a723ca62eb4f93f19c5c66beffaf70997afb64c
Fixes: commit 228d49bb842cda517b414bbfab460e8be429cc5f
Fixes: https://github.com/libguestfs/libguestfs/issues/184